Share WebView runtime command boundaries
