+++ crates/wrac_wxp_gui/Cargo.toml
@@ -14,6 +14,7 @@ log = "0.4"
 novonotes_run_loop = { workspace = true }
 parking_lot = "0.12"
 raw-window-handle = "0.6"
+serde = { version = "1", features = ["derive"] }
 serde_json = "1.0"
 url = "2"
 wrac_clap_adapter = { workspace = true }
