+++ plugins/wrac-gain/src-plugin/src/commands.rs
@@ -11,16 +11,14 @@ use serde_json::json;
 use wrac_clap_adapter::{
     HostContext, HostFamily, HostGuiResizeRequester, HostParamsEditNotifier, PluginDescriptor,
 };
-use wrac_wxp_gui::{WxpGuiResizeHandle, register_native_cursor_bridge_commands};
+use wrac_wxp_gui::{
+    WxpGuiResizeHandle, register_native_cursor_bridge_commands, register_resize_commands,
+};
 use wxp::{Channel, WxpCommandHandler};
 
 use crate::gui::{GuiStateNotifier, GuiSubscriptionId, editor_page_payload, parameter_payload};
 use crate::plugin::{parameter_default_value, parameter_host_value, parameter_text_value};
 use crate::state::{EditorPage, ProjectStateStore, SharedState};
-
-mod resize;
-
-use resize::register_resize_commands;
 
 pub(crate) struct CommandRegistrationDependencies {
     pub(crate) project_state: Arc<ProjectStateStore>,
@@ -249,6 +247,8 @@ pub(crate) fn register_commands(
         Ok::<_, String>(json!({ "ok": true }))
     });
 
+    // Window-management commands are shared wxp GUI plumbing, not WRAC Gain product state.
+    // Register them here so the product command list stays the single Rust/TS rendezvous point.
     register_resize_commands(
         &command_handler,
         host_gui_resize_requester,
